A bit of refactoring for static pages
